Fixed tests
Something is a little odd here as you have HasNulls as true, the ClrType information is not-nullable, but the array passed is nullable.
I think it worked before because nulls are removed so the newly created array is not nullable and won't contain nulls as they are removed.

Behavior now is the same as before.